[Cleanup] Update custom SFPI examples for macro wrappers
Move custom SFPI face functions into architecture-local SFPU headers and update the examples to use the non-templated SFPU macro wrappers so the docs and code follow the current LLK API.
